+++ frameworks/keyed/rendrjs/package.json
@@ -10,7 +10,7 @@
     "webpack-cli": "5.1.4"
   },
   "dependencies": {
-    "@rendrjs/core": "0.1.249"
+    "@rendrjs/core": "0.2.3"
   },
   "js-framework-benchmark": {
     "frameworkVersionFromPackage": "@rendrjs/core",
