--- frameworks/keyed/rendrjs/src/RemoveIcon.js
@@ -1,6 +1,0 @@
-import { span } from '@rendrjs/core';
-
-export let makeIcon = preload => span({
-    class: preload ? 'preloadicon ' : '' + 'glyphicon glyphicon-remove',
-    ariaHidden: true,
-});
